requirements.txt
git+https://github.com/BorealisAI/advertorch.git@cc8ccf971fbe28d0086d56defca2255d8f5d7b4e
attrs==21.4.0
boto3==1.21.3
botocore==1.24.3
brotlipy==0.7.0
certifi==2021.10.8
clearml==1.1.6
click==8.0.4
conda==4.10.3
conda-build==3.21.5
cycler==0.11.0
deepspeed==0.4.5
dnspython==2.1.0
docker-pycreds==0.4.0
fonttools==4.32.0
furl==2.1.3
future==0.18.2
fvcore==0.1.5.post20220212
gitdb==4.0.9
GitPython==3.1.26
importlib-metadata==4.11.1
importlib-resources==5.4.0
iopath==0.1.9
jmespath==0.10.0
jsonschema==4.4.0
kiwisolver==1.4.2
matplotlib==3.5.1
mkl-fft==1.3.1
mkl-service==2.4.0
ninja==1.10.2.3
olefile==0.46
orderedmultidict==1.0.1
packaging==21.3
pathlib2==2.3.7.post1
pathtools==0.1.2
Pillow==8.4.0
pkginfo==1.7.1
portalocker==2.4.0
promise==2.3
protobuf==3.19.4
pycosat==0.6.3
PyJWT==2.1.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-etcd==0.4.5
pytz==2021.3
PyYAML==5.4.1
s3transfer==0.5.1
scipy==1.7.3
sentry-sdk==1.5.5
shortuuid==1.0.8
smmap==5.0.0
tabulate==0.8.9
tensorboardX==1.8
termcolor==1.1.0
timm==0.5.4
torch==1.10.0
torchelastic==0.2.0
torchtext==0.11.0
torchvision==0.11.0
triton==1.1.1
wandb==0.12.10
yacs==0.1.8
yaspin==2.1.0
zipp==3.7.0